requirements.txt
absl-py==2.0.0
astor==0.8.1
dlib==19.20.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.10.0
importlib-metadata==6.8.0
imutils==0.5.4
Keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.5
numpy==1.26.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
protobuf==3.20.3
PyYAML==6.0.1
scipy==1.11.3
six==1.16.0
tensorboard==2.14.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
termcolor==2.3.0
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
fastapi==0.105.0
deepface==0.0.79
starlette==0.27.0
pymilvus==2.3.3
python-dotenv==1.0.0